+++ R/05-calculate_gi.R
@@ -40,8 +40,6 @@
 #' @param use_lfc Should Log fold change be used to calculate GI scores instead
 #' of CRISPR scores? If you do not have negative controls or CRISPR scores you
 #' will need to set this to TRUE.
-#' @param stats_by_rep Should statistics be calculated per rep or should replicates
-#' be collapsed?
 #' @return A gimap dataset with statistics and genetic interaction scores
 #' calculated. Overall results in the returned object can be obtained using
 #' gimap_dataset$overall_results Whereas target level genetic interaction
@@ -52,10 +50,13 @@
 #' @examples \donttest{
 #'
 #' gimap_dataset <- get_example_data("gimap",
-#'                                   data_dir = tempdir()) %>%
+#'   data_dir = tempdir()
+#' ) %>%
 #'   gimap_filter() %>%
-#'   gimap_annotate(cell_line = "HELA",
-#'                  annot_dir = tempdir()) %>%
+#'   gimap_annotate(
+#'     cell_line = "HELA",
+#'     annot_dir = tempdir()
+#'   ) %>%
 #'   gimap_normalize(
 #'     timepoints = "day",
 #'     missing_ids_file = tempfile()
@@ -66,8 +67,7 @@
 #' }
 calc_gi <- function(.data = NULL,
                     gimap_dataset,
-                    use_lfc = FALSE,
-                    stats_by_rep = FALSE) {
+                    use_lfc = FALSE) {
   # Summary the calculation
   # single_target_crispr_1 = geneA_nt1, geneA_nt2...
   # single_target_crispr_2 = nt1_geneB, nt2_geneB...
@@ -79,18 +79,22 @@ calc_gi <- function(.data = NULL,
   # expected_crispr_single_1=single_target_crispr_1 + mean_double_control_crispr
   # expected_crispr_single_2=single_target_crispr_2 + mean_double_control_crispr
 
-  # linear model is lm(observed_single_crispr ~ expected_single_crispr)
+  # linear model is with all samples:
+  # lm(mean_observed_single_crispr ~ mean_expected_single_crispr)
 
-  # single_target_gi_score = observed_single_crispr -
-  #      (intercept + slope * expected_single_crispr)
-  # double_target_gi_score = double_crispr_score -
-  #      (intercept + slope * expected_double_crispr)
+  # single_target_gi_score = mean_observed_single_crispr -
+  #      (intercept + slope * mean_expected_single_crispr)
+  # double_target_gi_score = mean_double_crispr_score -
+  #      (intercept + slope * mean_expected_double_crispr)
 
   # Code adapted from
   # https://github.com/FredHutch/GI_mapping/blob/main/workflow/scripts/
   # 04-calculate_GI_scores.Rmd
 
   if (!is.null(.data)) gimap_dataset <- .data
+
+  op <- options("R_MAX_VSIZE" = 32000000000)
+  on.exit(options(op))
 
   if (!("gimap_dataset" %in% class(gimap_dataset))) {
     stop(
@@ -110,8 +114,7 @@ calc_gi <- function(.data = NULL,
   lfc_adj <- gimap_dataset$normalized_log_fc
 
   if (!"crispr_score" %in% colnames(lfc_adj)) {
-
-   if (!use_lfc) {
+    if (!use_lfc) {
       stop("No CRISPR scores found, you can set use_lfc = TRUE to calculate
          genetic interaction scores using log fold change")
     }
@@ -120,6 +123,8 @@ calc_gi <- function(.data = NULL,
     lfc_adj$crispr_score <- lfc_adj$lfc
     message("Calculating Genetic Interaction scores using LFC")
   }
+
+  #### STEP 1 CALCULATE EXPECTED VALUES AND COLLAPSE TO MEANS
   # Get mean control target CRISPR scores -- they will be used for expected
   # calculations
   control_target_df <- lfc_adj %>%
@@ -130,7 +135,7 @@ calc_gi <- function(.data = NULL,
       values_to = "control_gRNA_seq"
     ) %>%
     # If there's the same control sequence, and rep
-    dplyr::group_by(rep, pgRNA_target, control_gRNA_seq, norm_ctrl_flag) %>%
+    dplyr::group_by(pgRNA_target, control_gRNA_seq, norm_ctrl_flag) %>%
     # Then take the mean for when controls have the same sequence
     dplyr::summarize(
       mean_double_control_crispr =
@@ -138,16 +143,14 @@ calc_gi <- function(.data = NULL,
       .groups = "drop"
     ) %>%
     dplyr::select(
-      rep,
       pgRNA_target,
       control_gRNA_seq,
       mean_double_control_crispr,
       norm_ctrl_flag,
     )
-  # This means we have a mean double control crispr for each rep and
-  # control sequence
+  # This means we have a mean double control crispr for each control sequence
 
-  # Calculate CRISPR scores for single targets
+  # Calculate expected and mean CRISPR scores for single targets
   single_crispr_df <- lfc_adj %>%
     dplyr::filter(target_type %in% c("ctrl_gene", "gene_ctrl")) %>%
     # We will be joining things based on the gRNA sequences so
@@ -167,47 +170,34 @@ calc_gi <- function(.data = NULL,
       ),
     ) %>%
     dplyr::left_join(control_target_df,
-      by = c("rep" = "rep", "control_gRNA_seq" = "control_gRNA_seq"),
+      by = c("control_gRNA_seq" = "control_gRNA_seq"),
       suffix = c("", "_control")
     ) %>%
-    group_by(rep, pgRNA_target, targeting_gRNA_seq) %>%
+    group_by(pgRNA_target, targeting_gRNA_seq) %>%
     # Taking the mean of the single target crisprs that have the same
     # targeting sequence
-    mutate(mean_single_crispr = mean(crispr_score, na.rm = TRUE)) %>%
-    dplyr::select(rep,
+    summarize(
+      mean_single_crispr = mean(crispr_score, na.rm = TRUE),
+      # Note this is now the mean of all the control sequences for a particular target
+      mean_double_control_crispr = mean(mean_double_control_crispr),
+      .groups = "drop"
+    ) %>%
+    dplyr::select(
       pgRNA_target,
       targeting_gRNA_seq,
-      control_gRNA_seq,
-      single_crispr = crispr_score,
       mean_single_crispr,
-      mean_double_control_crispr,
-      norm_ctrl_flag
+      mean_double_control_crispr
     ) %>%
-    dplyr::distinct()
+    ## calculate expected single-targeting GI score by summing the
+    ## single-targeting and the mean_double_control_crispr
+    dplyr::mutate(
+      expected_single_crispr = mean_single_crispr + mean_double_control_crispr,
+    )
+  # TODO: Ask alice do we want to combine target sequences as well?
 
-
-
-  if (nrow(control_target_df) > 0) {
-    single_crispr_df <- single_crispr_df %>%
-      ## calculate expected double-targeting GI score by summing the two mean
-      ## single-targeting
-      ## CRISPR scores for that paralog pair
-      dplyr::mutate(
-        expected_single_crispr = single_crispr + mean_double_control_crispr,
-      )
-  } else {
-    single_crispr_df <- single_crispr_df %>%
-      ## calculate expected double-targeting GI score by summing the two mean
-      ## single-targeting
-      ## CRISPR scores for that paralog pair
-      dplyr::mutate(
-        expected_single_crispr = single_crispr,
-      )
-  }
-
-  # Calculate expected
+  # This will be added on to the double crispr df
   expected_single_crispr_df <- single_crispr_df %>%
-    dplyr::select(rep, pgRNA_target, targeting_gRNA_seq, mean_single_crispr) %>%
+    dplyr::select(pgRNA_target, targeting_gRNA_seq, mean_single_crispr) %>%
     dplyr::distinct()
 
   # Now put it all together into one df
@@ -219,22 +209,19 @@ calc_gi <- function(.data = NULL,
       crispr_score,
       gRNA1_seq,
       gRNA2_seq,
-      pgRNA_target,
-      norm_ctrl_flag
+      pgRNA_target
     ) %>%
-    dplyr::distinct() %>%
     dplyr::left_join(expected_single_crispr_df,
-      by = c("rep" = "rep", "gRNA1_seq" = "targeting_gRNA_seq"),
+      by = c("gRNA1_seq" = "targeting_gRNA_seq"),
       suffix = c("", "_1")
     ) %>%
     dplyr::left_join(expected_single_crispr_df,
-      by = c("rep" = "rep", "gRNA2_seq" = "targeting_gRNA_seq"),
+      by = c("gRNA2_seq" = "targeting_gRNA_seq"),
       suffix = c("", "_2")
     ) %>%
-    dplyr::select(pg_ids,
-      rep,
+    dplyr::select(
+      pg_ids,
       double_crispr = crispr_score,
-      norm_ctrl_flag,
       gRNA1_seq,
       gRNA2_seq,
       pgRNA_target,
@@ -244,195 +231,113 @@ calc_gi <- function(.data = NULL,
     dplyr::mutate(
       expected_double_crispr = mean_single_crispr_1 + mean_single_crispr_2
     )
+  # TODO: Ask Alice: Should there be by reps at this point?
 
-  # Calculate the linear models from this
-  single_lm_df <- single_crispr_df %>%
-    dplyr::filter(!is.na(single_crispr), !is.na(expected_single_crispr)) %>%
-    dplyr::ungroup() %>%
-    dplyr::group_by(rep) %>%
-    group_modify(~ broom::tidy(
-      lm(single_crispr ~ expected_single_crispr, data = .x)
-    )) %>%
-    dplyr::select(term, estimate, rep) %>%
-    pivot_wider(
-      names_from = term,
-      values_from = estimate
-    ) %>%
-    rename(intercept = "(Intercept)", slope = expected_single_crispr)
+  #### STEP 2 LINEAR MODEL AND GI SCORE CALC
 
   message("Calculating Genetic Interaction scores")
 
+  # Calculate the linear models from this
+  # Previously we had a dplyr::group_by(rep) %>% here because we made linear models
+  # by each rep but we found doing a single linear model made more sense for
+  # expectation calculations
+  gimap_dataset$linear_model <- lm(mean_single_crispr ~ expected_single_crispr,
+    data = single_crispr_df
+  )
+
+  # Pull out the intercept and slope
+  calc_slope <- gimap_dataset$linear_model$coefficients[["expected_single_crispr"]]
+  calc_intercept <- gimap_dataset$linear_model$coefficients[["(Intercept)"]]
+
   # Do the linear model adjustments
   gi_calc_single <- single_crispr_df %>%
-    dplyr::left_join(single_lm_df, by = "rep") %>%
+    # Previously GI scores were calculated on a per construct basis but then we
+    # realized a mean_expected should be used because it is hypothetically closer
+    # to the expectation for the target overall.
+    dplyr::group_by(pgRNA_target) %>%
+    dplyr::summarize(
+      mean_expected_cs = mean(expected_single_crispr, na.rm = TRUE),
+      mean_observed_cs = mean(mean_single_crispr, na.rm = TRUE)
+    ) %>%
     dplyr::mutate(
-      single_gi_score = single_crispr -
-        (intercept + slope * expected_single_crispr)
+      gi_score = mean_observed_cs -
+        (calc_intercept + calc_slope * mean_expected_cs)
     )
 
   # Do the linear model adjustments but don't collapse double
   gi_calc_double <- double_crispr_df %>%
-    # Using the single target's linear model here
-    dplyr::left_join(single_lm_df, by = "rep") %>%
-    dplyr::mutate(
-      double_gi_score = double_crispr -
-        (intercept + slope * expected_double_crispr)
-    )
-
-  # Which replicates we have?
-  replicates <- unique(gi_calc_double$rep)
-
-  if (stats_by_rep) {
-    target_results <- lapply(replicates, function(replicate) {
-      gimap_stats(replicate,
-        gi_calc_single = gi_calc_single,
-        gi_calc_double = gi_calc_double
-      )
-    })
-
-    # Bring over the replicate names
-    names(target_results) <- replicates
-
-    # Turn into a data.frame
-    target_results_df <- dplyr::bind_rows(target_results, .id = "rep")
-
-    ## Clean up the data and get some means
-    gi_calc_double <- gi_calc_double %>%
-      dplyr::group_by(
-        rep,
-        pgRNA_target
-      )
-  } else {
-    target_results_df <- gimap_stats(
-      gi_calc_single = gi_calc_single,
-      gi_calc_double = gi_calc_double
-    )
-    ## Clean up the data and get some means
-    gi_calc_double <- gi_calc_double %>%
-      dplyr::group_by(
-        pgRNA_target
-      )
-  }
-
-  ## Clean up the data and get some means
-  gi_calc_double <- gi_calc_double %>%
+    dplyr::group_by(pg_ids, pgRNA_target) %>%
     dplyr::summarize(
       mean_expected_cs = mean(expected_double_crispr, na.rm = TRUE),
       mean_observed_cs = mean(double_crispr, na.rm = TRUE),
-      mean_gi_score = mean(double_gi_score, na.rm = TRUE)
+      .groups = "drop"
+    ) %>%
+    # Using the single target's linear model here
+    dplyr::mutate(
+      gi_score = mean_observed_cs -
+        (calc_intercept + calc_slope * mean_expected_cs)
+    )
+  # TODO: Check with Alice.
+  # We need at least more than one data point per target but currently have
+  # All 16 constructs going in where each is a mean of the reps
+
+
+  #### STEP 3: RUN THE TESTS
+  target_results_df <- gimap_stats(
+    gi_calc_single = gi_calc_single,
+    gi_calc_double = gi_calc_double
+  )
+
+  #### STEP 4: FORMATTING DATA FOR EASY STORING
+  gi_calc_double_clean <- gi_calc_double %>%
+    # Collapse to just stats and don't care about pg_ids anymore
+    dplyr::select(
+      pgRNA_target,
+      mean_expected_cs,
+      mean_observed_cs,
+      gi_score
+    ) %>%
+    dplyr::group_by(
+      pgRNA_target
+    ) %>%
+    # Collapse to one mean per target now that we've tested
+    dplyr::summarize(
+      mean_observed_cs = mean(mean_observed_cs, na.rm = TRUE),
+      mean_expected_cs = mean(mean_expected_cs, na.rm = TRUE),
+      gi_score = mean(gi_score, na.rm = TRUE),
+      target_type = "gene_gene"
     )
 
-  if (stats_by_rep) {
-    gi_calc_double <- gi_calc_double %>%
-      # Collapse to just stats and don't care about pg_ids anymore
-      dplyr::select(
-        rep,
-        pgRNA_target,
-        mean_expected_cs,
-        mean_observed_cs,
-        mean_gi_score
-      )
-  } else {
-    gi_calc_double <- gi_calc_double %>%
-      # Collapse to just stats and don't care about pg_ids anymore
-      dplyr::select(
-        pgRNA_target,
-        mean_expected_cs,
-        mean_observed_cs,
-        mean_gi_score
-      )
-  }
-  gi_calc_double <- gi_calc_double %>%
-    dplyr::mutate(target_type = "gene_gene") %>%
-    dplyr::distinct()
-
-  if (stats_by_rep) {
-    # Same kind of reformatting for single
-    gi_calc_single <- gi_calc_single %>%
-      dplyr::group_by(
-        rep,
-        pgRNA_target
-      )
-  } else {
-    # Same kind of reformatting for single
-    gi_calc_single <- gi_calc_single %>%
-      dplyr::group_by(
-        pgRNA_target
-      )
-  }
-  gi_calc_single <- gi_calc_single %>%
-    dplyr::summarize(
-      mean_expected_cs = mean(expected_single_crispr, na.rm = TRUE),
-      mean_observed_cs = mean(single_crispr, na.rm = TRUE),
-      mean_gi_score = mean(single_gi_score, na.rm = TRUE)
-    ) %>%
+  gi_calc_single_clean <- gi_calc_single %>%
     dplyr::mutate(target_type = dplyr::case_when(
       grepl("^ctrl_*", pgRNA_target) ~ "ctrl_gene",
       grepl("*_ctrl$", pgRNA_target) ~ "gene_ctrl"
-    ))
+    )) %>%
+    dplyr::select(
+      target_type,
+      pgRNA_target,
+      mean_expected_cs,
+      mean_observed_cs,
+      gi_score
+    ) %>%
+    dplyr::distinct()
 
-    if (stats_by_rep) {
-      gi_calc_single <- gi_calc_single %>%
-        dplyr::select(
-          rep,
-          target_type,
-          pgRNA_target,
-          mean_expected_cs,
-          mean_observed_cs,
-          mean_gi_score
-        ) %>%
-        dplyr::distinct()
-    } else {
-      gi_calc_single <- gi_calc_single %>%
-        dplyr::select(
-          target_type,
-          pgRNA_target,
-          mean_expected_cs,
-          mean_observed_cs,
-          mean_gi_score
-        ) %>%
-        dplyr::distinct()
-    }
+  all_gi_scores <- dplyr::bind_rows(gi_calc_double_clean, gi_calc_single_clean)
 
-  all_gi_scores <- dplyr::bind_rows(gi_calc_double, gi_calc_single)
-
-
-  if (stats_by_rep) {
   # Add on test results
   all_gi_scores <- all_gi_scores %>%
     dplyr::left_join(target_results_df,
-      by = c("pgRNA_target", "rep")
+      by = c("pgRNA_target")
     )
-  } else {
-    # Add on test results
-    all_gi_scores <- all_gi_scores %>%
-      dplyr::left_join(target_results_df,
-                       by = c("pgRNA_target")
-      )
-  }
 
   # Store the useful bits
   gimap_dataset$gi_scores <- all_gi_scores
-
-  # Store this
-  gimap_dataset$overall_results <- single_lm_df
-
-  means_by_rep <- dplyr::bind_rows(
-    dplyr::select(single_crispr_df,
-                  rep, pgRNA_target, norm_ctrl_flag, mean_score = mean_single_crispr,
-                  seq = targeting_gRNA_seq),
-    dplyr::select(double_crispr_df,
-                  rep, pgRNA_target, norm_ctrl_flag, mean_score = double_crispr),
-    dplyr::select(control_target_df,
-                  rep, pgRNA_target, norm_ctrl_flag, mean_score = mean_double_control_crispr,
-                  seq = control_gRNA_seq))
 
   if (!use_lfc) {
     # Save at the target level but call them crispr scores
     gimap_dataset$crispr_score$single_crispr_score <- single_crispr_df
     gimap_dataset$crispr_score$double_crispr_score <- double_crispr_df
     gimap_dataset$crispr_score$neg_control_crispr <- control_target_df
-    gimap_dataset$crispr_score$means_by_rep <- means_by_rep
   } else {
     # Save at the target level but they aren't crispr scores
     colnames(single_crispr_df) <- gsub("crispr", "lfc", colnames(single_crispr_df))
@@ -449,7 +354,6 @@ calc_gi <- function(.data = NULL,
         mean_expected_lfc = mean_expected_cs,
         mean_observed_lfc = mean_observed_cs
       )
-    gimap_dataset$crispr_score$means_by_rep <- means_by_rep
   }
 
   return(gimap_dataset)
@@ -469,19 +373,11 @@ gimap_stats <- function(gi_calc_double, gi_calc_single, replicate = NULL) {
   ## double-targeting GI scores for each paralog pair to the control vector
   ## adjust for multiple testing using the Benjamini-Hochberg method
 
-  if (!is.null(replicate)) {
-    gi_calc_double <- gi_calc_double %>%
-      dplyr::filter(rep == replicate)
-
-    gi_calc_single <- gi_calc_single %>%
-      dplyr::filter(rep == replicate)
-  }
-
   gi_scores <- gi_calc_double %>%
     group_by(pgRNA_target) %>%
     mutate(p_val = t.test(
-      x = gi_calc_single$single_gi_score,
-      y = double_gi_score, # all 16 construct guides here
+      x = gi_calc_single$gi_score,
+      y = gi_score, # all construct guides here
       paired = FALSE
     )$p.value)