+++ inst/doc/timepoint-example.R
@@ -11,14 +11,15 @@ library(gimap)
 library(dplyr)
 
 ## ----eval = FALSE-------------------------------------------------------------
-# example_data <- get_example_data("count")
+# example_data <- get_example_data("count") %>%
+#   dplyr::select(!Day05_RepA)
 
 ## ----eval = FALSE-------------------------------------------------------------
 # colnames(example_data)
 
 ## ----eval = FALSE-------------------------------------------------------------
 # counts <- example_data %>%
-#   select(c("Day00_RepA", "Day05_RepA", "Day22_RepA", "Day22_RepB", "Day22_RepC")) %>%
+#   select(c("Day00_RepA", "Day22_RepA", "Day22_RepB", "Day22_RepC")) %>%
 #   as.matrix()
 
 ## ----eval = FALSE-------------------------------------------------------------
@@ -27,9 +28,9 @@ library(dplyr)
 
 ## ----eval = FALSE-------------------------------------------------------------
 # sample_metadata <- data.frame(
-#   col_names = c("Day00_RepA", "Day05_RepA", "Day22_RepA", "Day22_RepB", "Day22_RepC"),
-#   day = as.numeric(c("0", "5", "22", "22", "22")),
-#   rep = as.factor(c("RepA", "RepA", "RepA", "RepB", "RepC"))
+#   col_names = c("Day00_RepA", "Day22_RepA", "Day22_RepB", "Day22_RepC"),
+#   day = as.numeric(c("0", "22", "22", "22")),
+#   rep = as.factor(c("RepA", "RepA", "RepB", "RepC"))
 # )
 
 ## ----eval = FALSE-------------------------------------------------------------
@@ -44,10 +45,11 @@ library(dplyr)
 
 ## ----eval = FALSE-------------------------------------------------------------
 # run_qc(gimap_dataset,
-#        output_file = "example_qc_report.Rmd",
-#        overwrite = TRUE,
-#        plots_dir = "plots",
-#        quiet = TRUE)
+#   output_file = file.path(output_dir, "example_qc_report.Rmd"),
+#   overwrite = TRUE,
+#   plots_dir = "plots",
+#   quiet = TRUE
+# )
 
 ## ----eval = FALSE-------------------------------------------------------------
 # gimap_dataset <- gimap_dataset %>%
@@ -74,19 +76,30 @@ library(dplyr)
 # head(dplyr::arrange(gimap_dataset$gi_score, fdr))
 
 ## ----eval = FALSE-------------------------------------------------------------
-# plot_exp_v_obs_scatter(gimap_dataset)
+# plot <- plot_exp_v_obs_scatter(gimap_dataset)
+# 
+# # Save it to a file
+# ggsave(file.path(output_dir, "exp_v_obs_scatter.png"))
 
 ## ----eval = FALSE-------------------------------------------------------------
 # plot_rank_scatter(gimap_dataset)
+# 
+# # Save it to a file
+# ggsave(file.path(output_dir, "plot_rank_scatter.png"))
 
 ## ----eval = FALSE-------------------------------------------------------------
 # plot_volcano(gimap_dataset)
+# 
+# # Save it to a file
+# ggsave(file.path(output_dir, "volcano_plot.png"))
 
 ## ----eval = FALSE-------------------------------------------------------------
-# # "NDEL1_NDE1" is top result so let's plot that
-# plot_targets_bar(gimap_dataset, target1 = "TIAL1", target2 = "TIA1",
-#                  reps_to_drop = "Day05_RepA_early")
+# # "CNOT8_CNOT7" is top result so let's plot that
+# plot_targets(gimap_dataset, target1 = "CNOT8", target2 = "CNOT7")
+# 
+# # Save it to a file
+# ggsave(file.path(output_dir, "CNOT8_CNOT7.png"))
 
-## ----eval = FALSE-------------------------------------------------------------
-# sessionInfo()
+## -----------------------------------------------------------------------------
+sessionInfo()