+++ tests/testthat/test-gi-calc.R
@@ -4,7 +4,8 @@ test_that("Test Genetic Interaction score calculations", {
     gimap_filter() %>%
     gimap_annotate(cell_line = "HELA") %>%
     gimap_normalize(
-      timepoints = "day"
+      timepoints = "day",
+      missing_ids_file = tempfile()
     ) %>%
     calc_gi()
 
@@ -14,31 +15,23 @@ test_that("Test Genetic Interaction score calculations", {
     slope = as.numeric(round(c(0.483, 0.659, 0.646, 0.650), 3))
   )
 
-  gimap_dataset$overall_results$intercept <- round(gimap_dataset$overall_results$intercept, 3)
-  gimap_dataset$overall_results$slope <- round(gimap_dataset$overall_results$slope, 3)
-
-  testthat::expect_true(all.equal(gimap_dataset$overall_results$rep, results$rep))
-  testthat::expect_identical(
-    round(gimap_dataset$overall_results$intercept[1], 3),
-    round(results$intercept[1], 3)
-  )
-  testthat::expect_identical(gimap_dataset$overall_results$rep[1], results$rep[1])
+  testthat::expect_type(gimap_dataset$linear_model, "list")
 
   testthat::expect_identical(
     round(gimap_dataset$gi_scores$mean_expected_cs[1], 3),
-    round(-0.2220, 3)
+    round(-0.4160, 3)
   )
   testthat::expect_identical(
     round(gimap_dataset$gi_scores$mean_observed_cs[1], 3),
-    round(-1.515, 3)
+    round(-0.5490, 3)
   )
   testthat::expect_identical(
-    round(gimap_dataset$gi_scores$mean_gi_score[1], 3),
-    round(-1.119, 3)
+    round(gimap_dataset$gi_scores$gi_score[1], 3),
+    round(-0.1470, 3)
   )
   testthat::expect_identical(
     round(gimap_dataset$gi_scores$p_val[1], 3),
-    round(0.001203125, 3)
+    round(0.2050, 3)
   )
 })
 
@@ -50,7 +43,8 @@ test_that("Test Genetic Interaction score calculations using LFC", {
     gimap_annotate(cell_line = "HELA") %>%
     gimap_normalize(
       timepoints = "day",
-      adj_method = "no_adjustment"
+      adj_method = "no_adjustment",
+      missing_ids_file = tempfile()
     ) %>%
     calc_gi(use_lfc = TRUE)
 
@@ -58,16 +52,46 @@ test_that("Test Genetic Interaction score calculations using LFC", {
 })
 
 
-test_that("Test Genetic Interaction score calculations by rep", {
+test_that("Test Genetic Interaction score without normalization", {
   testthat::skip_on_cran()
-  gimap_dataset <- get_example_data("gimap") %>%
+  gimap_dataset_wo <- get_example_data("gimap") %>%
     gimap_filter() %>%
     gimap_annotate(cell_line = "HELA") %>%
     gimap_normalize(
+      normalize_by_unexpressed = FALSE,
       timepoints = "day",
+      missing_ids_file = tempfile()
     ) %>%
-    calc_gi(stats_by_rep = TRUE)
+    calc_gi()
 
-  testthat::expect_true(class(gimap_dataset)[1] == "list")
+  gimap_dataset_w <- get_example_data("gimap") %>%
+    gimap_filter() %>%
+    gimap_annotate(cell_line = "HELA") %>%
+    gimap_normalize(
+      normalize_by_unexpressed = TRUE,
+      timepoints = "day",
+      missing_ids_file = tempfile()
+    ) %>%
+    calc_gi()
 
+
+  # Are the GI scores the same? No
+  testthat::expect_false(
+    all(gimap_dataset_wo$gi_scores$gi_score == gimap_dataset_w$gi_scores$gi_score)
+  )
+
+  # Are the log fold change values the same? No
+  testthat::expect_false(
+    all(gimap_dataset_wo$normalized_log_fc$lfc == gimap_dataset_w$normalized_log_fc$lfc)
+  )
+
+  # Are the CRISPR scores the same? No
+  testthat::expect_false(
+    all(gimap_dataset_wo$normalized_log_fc$crispr_score == gimap_dataset_w$normalized_log_fc$crispr_score)
+  )
+
+  table(
+    gimap_dataset_w$normalized_log_fc$unexpressed_ctrl_flag,
+    gimap_dataset_w$normalized_log_fc$norm_ctrl_flag
+  )
 })