+++ tests/testthat/test-run_qc.R
@@ -5,10 +5,11 @@ test_that("HTML file is created and content is correct", {
   gimap_dataset <- get_example_data("gimap")
 
   html_file <- run_qc(gimap_dataset,
-                      output_file = tempfile(),
-                      plots_dir = tempdir(),
-                      open_results = FALSE,
-                      overwrite = TRUE)
+    output_file = tempfile(),
+    plots_dir = tempdir(),
+    open_results = FALSE,
+    overwrite = TRUE
+  )
 
   expect_true(file.exists(html_file))
 
@@ -16,10 +17,11 @@ test_that("HTML file is created and content is correct", {
   gimap_dataset <- get_example_data("gimap_treatment")
 
   html_file <- run_qc(gimap_dataset,
-                      output_file = tempfile(),
-                      plots_dir = tempdir(),
-                      open_results = FALSE,
-                      overwrite = TRUE)
+    output_file = tempfile(),
+    plots_dir = tempdir(),
+    open_results = FALSE,
+    overwrite = TRUE
+  )
 
   expect_true(file.exists(html_file))
 })