+++ NEWS.md
@@ -1,3 +1,13 @@
+
+
+# gimap 1.1.1
+
+* Downloads example data from figshare instead of storing in the package itself due to CRAN size requirements. 
+
+# gimap 1.1.0
+
+* Adjustments to gimap calculations -- all reps collapsed and linear models done with mean across reps.
+
 # gimap 1.0.3
 
 * Polish how annotation data is saved